ffmpeg: vaapi: fix support for cropped frames.
FFmpeg does not expose any API to determine the cropping region. It
is assumed that the SW decoded pixels data is already offset to fit
the cropped region only. i.e. libavcodec alters AVFrame.data[0], and
subsequent planes, to fit the cropped region. So, this means that we
can reconstruct the cropping rectangle by detecting changes to data[0],
as this should still represent the VA surface id (data[3]).