gstreamer: make it possible to filter out video parser caps.
Make it possible to filter out video parser caps on the src pad as
if the selection was made from the decode element. This uses the
"autoplug-query" signal and the new per-module option system.

For instance:
-o vparse_capsfilter=video/x-h264,stream-format=byte-stream,alignment=nal
will make it possible to filter out the output caps to use for the
h264parse element.