gstreamer: completely disable vaapidecodebin element for now.
Disable the "vaapidecodebin" element for now since it would automatically
insert a vaapipostproc element, thus performing deinterlacing whenever it
is necessary. Since we are only interested in validating the output as it
comes out of the decoder, keeping the vaapidecodebin element to insert a
deinterlacer is going to perturb the process.